refactor: split process-wide engine setup into TUnitProcessInitializer
Addresses PR review finding 4 (static guard + instance dependency
mismatch in TUnitInitializer). Lift process-once concerns — global
exception handlers, the throw-on-trace-assert listener, and test
parameter parsing — out of the per-session TUnitInitializer into a
dedicated static type whose name makes the lifetime obvious.

TUnitInitializer now does only session-scoped work and delegates
explicit ICommandLineOptions to the process initializer. The static
flag for parameter parsing is gone; the per-instance guard left in
TUnitInitializer protects only concurrent ExecuteRequestAsync calls
within a single session, which is its actual job.
